backend/deepFakeDetection/requirements.txt
# Deep Learning
torch
torchvision
torchaudio

# Computer Vision
opencv-python
Pillow
numpy

# Data Handling
pandas
scikit-learn
tqdm

# Face Detection
facenet-pytorch

# API Server
fastapi
uvicorn
python-multipart

# Utilities
matplotlib
seaborn
timm